Split spec in every seed
