+++ app/models/domicile_seed.rb
@@ -1,3 +1,7 @@
 class DomicileSeed < DomicileBase
   include Garden::Seed
+
+  before_create do
+    self.country = country.upcase
+  end
 end
