+++ app/models/identification_seed.rb
@@ -1,3 +1,7 @@
 class IdentificationSeed < IdentificationBase
   include Garden::Seed
+
+  before_create do
+    self.issuer = issuer.upcase
+  end
 end
