+++ app/models/legal_entity_docket_seed.rb
@@ -1,3 +1,7 @@
 class LegalEntityDocketSeed < LegalEntityDocketBase
   include Garden::Seed
+
+  before_create do
+    self.country = country.upcase
+  end
 end
