+++ app/models/phone_seed.rb
@@ -1,3 +1,7 @@
 class PhoneSeed < PhoneBase
   include Garden::Seed
+
+  before_create do
+    self.country = country.upcase
+  end
 end
