+++ spec/models/identification_seed_spec.rb
@@ -41,4 +41,17 @@ describe IdentificationSeed do
   it 'is valid with an issue' do
     expect(valid_seed).to be_valid
   end
+
+  it 'sets issuer to upper case' do
+    seed = IdentificationSeed.new(
+      number: '2545566',
+      identification_kind_code: 'national_id',
+      issuer: 'mx',
+      issue: create(:basic_issue)
+    )
+
+    seed.save
+
+    expect(seed.issuer).to eq('MX')
+  end
 end
